[ALPS07240294] Fix coverity issue
10888391 / change variable type from int to long

MTK-Commit-Id: aa7bb138fb60640d904e8b01f3d7e6d4bf1b972e

CR-Id: ALPS07240294
Feature: Thermal Management

+++ drivers/thermal/mediatek/thermal_trace.h
@@ -61,7 +61,7 @@ TRACE_EVENT(md_tx_pwr_limit,
 	TP_ARGS(md_cdev, status),
 
 	TP_STRUCT__entry(
-		__field(unsigned int, state)
+		__field(unsigned long, state)
 		__field(unsigned int, pwr)
 		__field(unsigned int, id)
 		__field(enum md_cooling_status, status)
